We previously had a bug where we were maybe-using the wrong resid_post?
The hook was definitely broken, I think we were overwriting the
"unembed" key of the rotated acts with the unrotated ones so we only
broke the case of trying to do feature viz on pre-unembed. But this
change also allowed us to pass the proper unsliced 65d into the viz
code? Maybe previously we were only messing up the shape? Anyway this
should be correct now but treat with caution.